Chunk out duplicate issue checker
