+++ lib/open-issues.js
@@ -5,6 +5,7 @@ const organizeCommits = require('./organize-commits')
 const generateLabel = require('./generate-label')
 const reopenClosed = require('./reopen-closed')
 const metadata = require('./metadata')
+const checkForDuplicate = require('./check-for-duplicate-issue')
 const {assignFlow, getAppComments} = require('./helpers')
 
 module.exports = async (context, robot) => {
@@ -45,32 +46,17 @@ module.exports = async (context, robot) => {
 
       titles.forEach(async title => {
         // Check if an issue with that title exists
-        const search = await context.github.search.issues({
-          q: `${title} in:title repo:${context.payload.repository.full_name}`,
-          per_page: 100
-        })
-
-        if (search.data.total_count !== 0) {
-          const existingIssue = search.data.items.find(issue => {
-            if (!issue.body) return false
-            const titleKey = metadata(context, issue).get('title')
-            const fileKey = metadata(context, issue).get('file')
-            return titleKey === title && fileKey === file
-          })
-
-          if (existingIssue) {
-            if (cfg.reopenClosed && existingIssue.state === 'closed') {
-              return reopenClosed(robot.log, context, cfg, existingIssue.number, file, commitSha)
-            } else {
-              return
-            }
-          }
+        const existingIssue = await checkForDuplicate(context, cfg, title, file)
+        if (existingIssue) {
+          return reopenClosed(robot.log, context, cfg, existingIssue.number, file, commitSha)
+        } else if (existingIssue === null) {
+          return
         }
 
         const body = generateBody(context, cfg, title, file, contents, author, commitSha, pr)
 
         if (pr) {
-          const todoComments = await getAppComments(pr)
+          const todoComments = await getAppComments(context, pr)
 
           // Get comments on the pull request (via the issues API)
           const existingPRComment = todoComments.some(comment => {
